renamed project, removed number from class name
