+++ code/binary-search/View/src/main/java/ru/unn/agile/binarysearch/view/Search.java
@@ -1,0 +1,9 @@
+package ru.unn.agile.binarysearch.view;
+
+import javafx.fxml.FXML;
+
+public class Search {
+    @FXML
+    void initialize() {
+    }
+}
