+++ code/binary-search/ViewModel/src/main/java/ru/unn/agile/binarysearch/viewmodel/ViewModel.java
@@ -1,0 +1,16 @@
+package ru.unn.agile.binarysearch.viewmodel;
+
+import javafx.beans.binding.BooleanBinding;
+import javafx.beans.property.*;
+import javafx.beans.value.ChangeListener;
+import javafx.beans.value.ObservableValue;
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
+import ru.unn.agile.binarysearch.model.BinarySearch;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class ViewModel {
+
+}
