+++ code/binary-search/ViewModel/src/test/java/ru/unn/agile/binarysearch/viewmodel/ViewModelTests.java
@@ -1,0 +1,12 @@
+package ru.unn.agile.binarysearch.viewmodel;
+
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+import ru.unn.agile.binarysearch.model.BinarySearch;
+
+import static org.junit.Assert.*;
+
+public class ViewModelTests {
+
+}
