--- code/gladkova-tatiana-binary-search/View/src/main/java/ru/unn/agile/number/binarysearch/view/Search.java
@@ -1,9 +1,0 @@
-package ru.unn.agile.number.binarysearch.view;
-
-import javafx.fxml.FXML;
-
-public class Search {
-    @FXML
-    void initialize() {
-    }
-}
