+++ settings.gradle
@@ -131,10 +131,12 @@ include "code/salary-calculator"
 project(":code/salary-calculator").name = "salary-calculator"
 
 // Gladkova Tatiana, Binary Search
-include "code/gladkova-tatiana-binary-search/Model"
-project(":code/gladkova-tatiana-binary-search/Model").name = "gladkova-tatiana-binary-search"
-include "code/gladkova-tatiana-binary-search/View"
-project(":code/gladkova-tatiana-binary-search/View").name = "gladkova-tatiana-binary-search-View"
+include "code/binary-search/Model"
+project(":code/binary-search/Model").name = "binary-search"
+include "code/binary-search/View"
+project(":code/binary-search/View").name = "binary-search-View"
+include "code/binary-search/ViewModel"
+project(":code/binary-search/ViewModel").name = "binary-search-ViewModel"
 
 // Kryukova Polina, caesar-cipher
 include "code/caesar-cipher"
